centralizando todos os metodos do google na main.py
