--- autenticadores.py
@@ -1,38 +1,0 @@
-
-from authlib.client import AssertionSession
-import json
-
-
-
-
-
-def google_api_auth(arqv_json='credenciais/colaborabot-gAPI.json', subject=None):
-    with open(arqv_json, 'r') as f:
-        conf = json.load(f)
-
-    token_url = conf['token_uri']
-    issuer = conf['client_email']
-    key = conf['private_key']
-    key_id = conf.get('private_key_id')
-
-    header = {'alg': 'RS256'}
-    scopes = [
-        'https://spreadsheets.google.com/feeds',
-        'https://www.googleapis.com/auth/drive'
-    ]
-
-    if key_id:
-        header['kid'] = key_id
-
-    # Google puts scope in payload
-    claims = {'scope': ' '.join(scopes)}
-    return AssertionSession(
-        grant_type=AssertionSession.JWT_BEARER_GRANT_TYPE,
-        token_url=token_url,
-        issuer=issuer,
-        audience=token_url,
-        claims=claims,
-        subject=subject,
-        key=key,
-        header=header,
-    )
