+++ main.py
@@ -2,13 +2,10 @@
 from time import sleep
 
 from colaborabot import Colaborabot
+import gspread
+from authlib.client import AssertionSession
+import json
 
-# não é utilizado??
-# def criar_tweet(url, orgao):
-#     """
-#     Criando o tweet com o status do site recém acessado
-#     """
-#     twitter_bot.update_status(lista_frases(url=url, orgao=orgao))
 
 # Guardando informações de hora e data da máquina
 
@@ -62,6 +59,45 @@ def plan_csv():
         pasta_logs.mkdir()
     arq_log.write_text(data='data_bsb;data_utc;url;orgao;cod_resposta\n', encoding='UTF8')
     return arq_log
+
+def google_sshet():
+    """
+    Função simples para retornar um objeto capaz de manipular as planilhas do Google Sheets.
+    """
+    session = google_api_auth()
+    ggle_cred = gspread.Client(None, session)
+    return ggle_cred
+
+def google_api_auth(arqv_json='credenciais/colaborabot-gAPI.json', subject=None):
+    with open(arqv_json, 'r') as f:
+        conf = json.load(f)
+
+    token_url = conf['token_uri']
+    issuer = conf['client_email']
+    key = conf['private_key']
+    key_id = conf.get('private_key_id')
+
+    header = {'alg': 'RS256'}
+    scopes = [
+        'https://spreadsheets.google.com/feeds',
+        'https://www.googleapis.com/auth/drive'
+    ]
+
+    if key_id:
+        header['kid'] = key_id
+
+    # Google puts scope in payload
+    claims = {'scope': ' '.join(scopes)}
+    return AssertionSession(
+        grant_type=AssertionSession.JWT_BEARER_GRANT_TYPE,
+        token_url=token_url,
+        issuer=issuer,
+        audience=token_url,
+        claims=claims,
+        subject=subject,
+        key=key,
+        header=header,
+    )
